refactor: re-organized files and extracted components to be re-used
