+++ src/components/organisms/storage/listing/BaseSettings.tsx
@@ -7,7 +7,7 @@ import MenuItem from '@material-ui/core/MenuItem'
 import InputAdornment from '@material-ui/core/InputAdornment'
 import Tooltip from '@material-ui/core/Tooltip'
 import { colors, validatedNumber } from '@rsksmart/rif-ui'
-import CountrySelect from 'components/molecules/storage/CountrySelect'
+import CountrySelect from 'components/molecules/CountrySelect'
 import StorageListingStore from 'store/Market/storage/ListingStore'
 import { CountryType } from 'models/Country'
 import { StorageListingStoreProps } from 'store/Market/storage/interfaces'
